+++ src/test/java/org/dinosaur/foodbowl/domain/review/persistence/ReviewCustomRepositoryTest.java
@@ -1,15 +1,12 @@
 package org.dinosaur.foodbowl.domain.review.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.math.BigDecimal;
 import java.util.List;
 import org.dinosaur.foodbowl.domain.member.domain.Member;
-import org.dinosaur.foodbowl.domain.photo.domain.Photo;
 import org.dinosaur.foodbowl.domain.review.application.dto.CoordinateBoundDto;
 import org.dinosaur.foodbowl.domain.review.domain.Review;
-import org.dinosaur.foodbowl.domain.review.domain.ReviewPhoto;
 import org.dinosaur.foodbowl.domain.store.domain.Store;
 import org.dinosaur.foodbowl.test.PersistenceTest;
 import org.junit.jupiter.api.Nested;
@@ -17,13 +14,10 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
 @SuppressWarnings("NonAsciiCharacters")
-class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
+class ReviewCustomRepositoryTest extends PersistenceTest {
 
     @Autowired
-    private ReviewPhotoRepository reviewPhotoRepository;
-
-    @Autowired
-    private ReviewPhotoCustomRepositoryImpl reviewPhotoCustomRepositoryImpl;
+    private ReviewCustomRepositoryImpl reviewCustomRepository;
 
     @Nested
     class 팔로잉_유저의_리뷰_목록_페이징_조회_시 {
@@ -42,7 +36,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     review.getId() + 1,
                     coordinateBoundDto,
@@ -66,7 +60,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     review.getId() - 1,
                     coordinateBoundDto,
@@ -90,7 +84,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -113,7 +107,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -137,7 +131,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(1)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -161,7 +155,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(1)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -185,7 +179,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(1)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -210,7 +204,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -236,7 +230,7 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
                     BigDecimal.valueOf(3)
             );
 
-            List<Review> result = reviewPhotoCustomRepositoryImpl.getPaginationReviewsByFollowing(
+            List<Review> result = reviewCustomRepository.getPaginationReviewsByFollowing(
                     member.getId(),
                     null,
                     coordinateBoundDto,
@@ -245,59 +239,5 @@ class ReviewPhotoCustomRepositoryImplTest extends PersistenceTest {
 
             assertThat(result).containsExactly(reviewC, reviewB);
         }
-    }
-
-    @Test
-    void 리뷰_사진_엔티티를_삭제한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo photo1 = photoTestPersister.builder().save();
-        Photo photo2 = photoTestPersister.builder().save();
-        ReviewPhoto reviewPhoto1 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo1)
-                .build();
-        ReviewPhoto reviewPhoto2 = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo2)
-                .build();
-        reviewPhotoRepository.save(reviewPhoto1);
-        reviewPhotoRepository.save(reviewPhoto2);
-
-        long deleteCount = reviewPhotoCustomRepositoryImpl.deleteAllByReview(review);
-
-        assertThat(deleteCount).isEqualTo(2);
-    }
-
-    @Test
-    void 리뷰에_해당하는_사진_엔티티들을_삭제한다() {
-        Review review = reviewTestPersister.builder().save();
-        Photo deleteTargetPhotoA = photoTestPersister.builder().save();
-        Photo deleteTargetPhotoB = photoTestPersister.builder().save();
-        Photo photo = photoTestPersister.builder().save();
-        ReviewPhoto deleteReviewPhotoA = ReviewPhoto.builder()
-                .review(review)
-                .photo(deleteTargetPhotoA)
-                .build();
-        ReviewPhoto deleteReviewPhotoB = ReviewPhoto.builder()
-                .review(review)
-                .photo(deleteTargetPhotoB)
-                .build();
-        ReviewPhoto reviewPhoto = ReviewPhoto.builder()
-                .review(review)
-                .photo(photo)
-                .build();
-        reviewPhotoRepository.save(deleteReviewPhotoA);
-        reviewPhotoRepository.save(deleteReviewPhotoB);
-        reviewPhotoRepository.save(reviewPhoto);
-
-        List<Photo> deleteTargetPhotos = List.of(deleteTargetPhotoA, deleteTargetPhotoB);
-        long deleteCount = reviewPhotoCustomRepositoryImpl.deleteByReviewAndPhotos(review, deleteTargetPhotos);
-
-        assertSoftly(softly -> {
-            softly.assertThat(deleteCount).isEqualTo(deleteTargetPhotos.size());
-            softly.assertThat(reviewPhotoRepository.findAllByReview(review)).containsExactly(reviewPhoto);
-            softly.assertThat(reviewPhotoRepository.findAllByReview(review))
-                    .doesNotContain(deleteReviewPhotoA, deleteReviewPhotoB);
-        });
     }
 }
